+++ NEWS.md
@@ -8,13 +8,13 @@
     + check() produces 2 'notes'
   
 * 0.9.6
-    + Initial 'open beta' relase
     + Changed 'Modularity_Signed' to 'signedModularity'
     + Revised signedModularity
     + Changed S4 naming convention from 'rnet.basic' to 'rnetBasic'
     + `check()` produces 0 errors/warnings/notes (other than UTF-8 warning)
 
 * 0.9.7
-    + Removed signedModularity.data.frame due to flawed algorithm
-    + Revised signedModularity to matrix-based algorith instead of list-based
+    + Initial 'open beta' relase
+    + Removed signedModularity.data.frame 
+    + Revised signedModularity to matrix-based algorithm instead of list-based